requirements.txt
altgraph==0.17.3
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyserial==3.5
python-gsmmodem-new==0.13.0
requests==2.28.2
SQLAlchemy==2.0.10
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
